tests/requirements-dev-minimal.txt
boto3>=1.9
feather-format>=0.4.1
freezegun>=0.3.15
google-cloud-storage>=1.28.0
mock>=3.0.5
moto>=1.3.7
pyarrow>=0.12.0
pytest>=5.3.5
pytest-cov>=2.8.1
requirements-parser>=0.2.0
xlrd>=1.1.0
